requirements.txt
chatterbox==0.0.0
ffmpeg==1.4
google_api_python_client==2.179.0
google_auth_oauthlib==1.2.2
protobuf==6.32.0
python-dotenv==1.1.1
torch==2.8.0
torchaudio==2.8.0
whisper==1.1.10